+++ core/api/src/main/java/com/blazebit/persistence/spi/JpqlFunctionKind.java
@@ -28,6 +28,10 @@ public enum JpqlFunctionKind {
      */
     AGGREGATE,
     /**
+     * An ordered set aggregate function.
+     */
+    ORDERED_SET_AGGREGATE,
+    /**
      * A window function.
      */
     WINDOW,