+++ core/impl/src/main/java/com/blazebit/persistence/impl/CriteriaBuilderFactoryImpl.java
@@ -163,6 +163,8 @@ public class CriteriaBuilderFactoryImpl implements CriteriaBuilderFactory {
                 return FunctionKind.VOLATILE;
             case AGGREGATE:
                 return FunctionKind.AGGREGATE;
+            case ORDERED_SET_AGGREGATE:
+                return FunctionKind.ORDERED_SET_AGGREGATE;
             case DETERMINISTIC:
                 return FunctionKind.DETERMINISTIC;
             default: