+++ core/parser/src/main/java/com/blazebit/persistence/parser/FunctionKind.java
@@ -23,6 +23,7 @@ package com.blazebit.persistence.parser;
  */
 public enum FunctionKind {
     AGGREGATE,
+    ORDERED_SET_AGGREGATE,
     WINDOW,
     VOLATILE,
     DETERMINISTIC;