+++ core/parser/src/main/java/com/blazebit/persistence/parser/expression/JPQLNextExpressionVisitorImpl.java
@@ -401,7 +401,7 @@ public class JPQLNextExpressionVisitorImpl extends JPQLNextParserBaseVisitor<Exp
         } else {
             arguments = Collections.emptyList();
         }
-        return handleFunction("COUNT", distinct, arguments, ctx, ctx.whereClause(), ctx.windowName, ctx.windowDefinition());
+        return handleFunction("COUNT", distinct, arguments, ctx, null, ctx.whereClause(), ctx.windowName, ctx.windowDefinition());
     }
 
     @Override
@@ -413,10 +413,10 @@ public class JPQLNextExpressionVisitorImpl extends JPQLNextParserBaseVisitor<Exp
         for (int i = 0; i < size; i++) {
             arguments.add(expressions.get(i).accept(this));
         }
-        return handleFunction(ctx.name.getText(), distinct, arguments, ctx, ctx.whereClause(), ctx.windowName, ctx.windowDefinition());
+        return handleFunction(ctx.name.getText(), distinct, arguments, ctx, ctx.orderByClause(), ctx.whereClause(), ctx.windowName, ctx.windowDefinition());
     }
 
-    private Expression handleFunction(String name, boolean distinct, List<Expression> arguments, ParserRuleContext ctx, JPQLNextParser.WhereClauseContext whereClauseContext, JPQLNextParser.IdentifierContext windowName, JPQLNextParser.WindowDefinitionContext windowDefinitionContext) {
+    private Expression handleFunction(String name, boolean distinct, List<Expression> arguments, ParserRuleContext ctx, JPQLNextParser.OrderByClauseContext orderByClauseContext, JPQLNextParser.WhereClauseContext whereClauseContext, JPQLNextParser.IdentifierContext windowName, JPQLNextParser.WindowDefinitionContext windowDefinitionContext) {
         String lowerName = name.toLowerCase();
         FunctionKind functionKind;
         // Builtin functions
@@ -425,7 +425,7 @@ public class JPQLNextExpressionVisitorImpl extends JPQLNextParserBaseVisitor<Exp
             case "current_date":
             case "current_time":
             case "current_timestamp":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 return new FunctionExpression(name, Collections.<Expression>emptyList());
             case "outer":
                 if (!allowOuter) {
@@ -442,40 +442,40 @@ public class JPQLNextExpressionVisitorImpl extends JPQLNextParserBaseVisitor<Exp
             case "mod":
             case "coalesce":
             case "nullif":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 return new FunctionExpression(name, arguments);
             case "trim":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 return new TrimExpression(Trimspec.BOTH, null, arguments.get(0));
             case "size":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 ((PathExpression) arguments.get(0)).setUsedInCollectionFunction(true);
                 return new FunctionExpression(name, arguments);
             case "index":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 PathExpression listIndexPath = (PathExpression) arguments.get(0);
                 listIndexPath.setCollectionQualifiedPath(true);
                 return new ListIndexExpression(listIndexPath);
             case "key":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 PathExpression mapKeyPath = (PathExpression) arguments.get(0);
                 mapKeyPath.setCollectionQualifiedPath(true);
                 return new MapKeyExpression(mapKeyPath);
             case "value":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 PathExpression mapValuePath = (PathExpression) arguments.get(0);
                 mapValuePath.setCollectionQualifiedPath(true);
                 return new MapValueExpression(mapValuePath);
             case "entry":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 PathExpression mapEntryPath = (PathExpression) arguments.get(0);
                 mapEntryPath.setCollectionQualifiedPath(true);
                 return new MapEntryExpression(mapEntryPath);
             case "type":
-                failDistinct(distinct, ctx);
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
                 return new TypeFunctionExpression(arguments.get(0));
             case "function":
-                failDistinct(distinct, ctx);
+                assertNotDistinct(distinct, ctx);
                 String functionName = ((StringLiteral) arguments.get(0)).getValue();
                 functionKind = functions.get(functionName.toLowerCase());
                 if (functionKind == null) {
@@ -489,27 +489,52 @@ public class JPQLNextExpressionVisitorImpl extends JPQLNextParserBaseVisitor<Exp
             //CHECKSTYLE:ON: FallThrough
         }
         if (functionKind == null) {
-            if (whereClauseContext == null && windowName == null && windowDefinitionContext == null) {
-                failDistinct(distinct, ctx);
-                return handleMacro(name, arguments, ctx);
-            }
-            throw new SyntaxErrorException("No function with the name '" + name + "' exists!");
+            assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
+            return handleMacro(name, arguments, ctx);
         }
-        if (functionKind == FunctionKind.AGGREGATE) {
-            // NOTE: We currently don't support JUST filtering for aggregate functions, but maybe in the future
-            if (windowName == null && windowDefinitionContext == null) {
-                if (whereClauseContext == null) {
-                    return new AggregateExpression(distinct, name, arguments);
-                } else {
-                    return new AggregateExpression(distinct, "window_" + name, arguments, (Predicate) whereClauseContext.predicate().accept(this));
+        switch (functionKind) {
+            case AGGREGATE:
+                if (orderByClauseContext != null) {
+                    throw new SyntaxErrorException("Invalid use of WITHIN GROUP for function: " + getInputText(ctx));
                 }
-            } else {
-                failDistinct(distinct, ctx);
-                return new FunctionExpression("window_" + name, arguments, createWindowDefinition(whereClauseContext, windowName, windowDefinitionContext, functionKind));
-            }
-        } else {
-            failDistinct(distinct, ctx);
-            return new FunctionExpression(name, arguments, createWindowDefinition(whereClauseContext, windowName, windowDefinitionContext, functionKind));
+                // NOTE: We currently don't support JUST filtering for aggregate functions, but maybe in the future
+                if (windowName == null && windowDefinitionContext == null) {
+                    if (whereClauseContext == null) {
+                        return new AggregateExpression(distinct, name, arguments);
+                    } else {
+                        return new AggregateExpression(distinct, "window_" + name, arguments, null, (Predicate) whereClauseContext.predicate().accept(this));
+                    }
+                } else {
+                    if (distinct) {
+                        arguments.add(0, new StringLiteral("DISTINCT"));
+                    }
+                    return new FunctionExpression("window_" + name, arguments, null, createWindowDefinition(whereClauseContext, windowName, windowDefinitionContext, functionKind));
+                }
+            case ORDERED_SET_AGGREGATE:
+                // NOTE: We currently don't support JUST filtering for aggregate functions, but maybe in the future
+                if (windowName == null && windowDefinitionContext == null) {
+                    if (orderByClauseContext == null && whereClauseContext == null) {
+                        return new AggregateExpression(distinct, name, arguments);
+                    } else {
+                        return new AggregateExpression(distinct, name, arguments, createOrderByItems(orderByClauseContext), whereClauseContext == null ? null : (Predicate) whereClauseContext.predicate().accept(this));
+                    }
+                } else {
+                    if (distinct) {
+                        arguments.add(0, new StringLiteral("DISTINCT"));
+                    }
+                    return new FunctionExpression(name, arguments, createOrderByItems(orderByClauseContext), createWindowDefinition(whereClauseContext, windowName, windowDefinitionContext, functionKind));
+                }
+            case WINDOW:
+                if (orderByClauseContext != null) {
+                    throw new SyntaxErrorException("Invalid use of WITHIN GROUP for function: " + getInputText(ctx));
+                }
+                if (distinct) {
+                    arguments.add(0, new StringLiteral("DISTINCT"));
+                }
+                return new FunctionExpression(name, arguments, null, createWindowDefinition(whereClauseContext, windowName, windowDefinitionContext, functionKind));
+            default:
+                assertNormalFunctionInvocation(distinct, ctx, orderByClauseContext, whereClauseContext, windowName, windowDefinitionContext);
+                return new FunctionExpression(name, arguments, null, null);
         }
     }
 
@@ -529,7 +554,20 @@ public class JPQLNextExpressionVisitorImpl extends JPQLNextParserBaseVisitor<Exp
         }
     }
 
-    private void failDistinct(boolean distinct, ParserRuleContext ctx) {
+    private void assertNormalFunctionInvocation(boolean distinct, ParserRuleContext ctx, JPQLNextParser.OrderByClauseContext orderByClauseContext, JPQLNextParser.WhereClauseContext whereClauseContext, JPQLNextParser.IdentifierContext windowName, JPQLNextParser.WindowDefinitionContext windowDefinitionContext) {
+        assertNotDistinct(distinct, ctx);
+        if (orderByClauseContext != null) {
+            throw new SyntaxErrorException("Invalid use of WITHIN GROUP for function: " + getInputText(ctx));
+        }
+        if (whereClauseContext != null) {
+            throw new SyntaxErrorException("Invalid use of FILTER for function: " + getInputText(ctx));
+        }
+        if (windowName != null || windowDefinitionContext != null) {
+            throw new SyntaxErrorException("Invalid use of OVER for function: " + getInputText(ctx));
+        }
+    }
+
+    private void assertNotDistinct(boolean distinct, ParserRuleContext ctx) {
         if (distinct) {
             throw new SyntaxErrorException("Invalid use of DISTINCT for function: " + getInputText(ctx));
         }